updatecart

+++ controller/cartCURD.php
@@ -1,5 +1,5 @@
 <?php
-
+include './model/config.php';
 session_start();
 
 $user_id = $_SESSION['user_id'];
@@ -8,31 +8,18 @@ if (!isset($user_id)) {
     header('location:login.php');
 }
 
-if (isset($_POST['update_cart'])) {
-    $cart_id = $_POST['cart_id'];
-    $cart_quantity = $_POST['cart_quantity'];
-    
-    $update_cart_query = $pdo->prepare("UPDATE `cart` SET quantity = :cart_quantity WHERE id = :cart_id");
-    $update_cart_query->bindParam(':cart_quantity', $cart_quantity, PDO::PARAM_INT);
-    $update_cart_query->bindParam(':cart_id', $cart_id, PDO::PARAM_INT);
-    
-    if ($update_cart_query->execute()) {
-        $message[] = 'Số lượng sản phẩm đã được cập nhật!';
-    } else {
-        $message[] = 'Không thể cập nhật số lượng!';
-    }
-}
-
 if (isset($_GET['delete'])) {
     $delete_id = $_GET['delete'];
 
-    $delete_cart_query = $pdo->prepare("DELETE FROM `cart` WHERE id = :delete_id");
+    $delete_cart_query = $pdo->prepare("DELETE FROM `cart` WHERE id = :delete_id AND user_id = :user_id");
     $delete_cart_query->bindParam(':delete_id', $delete_id, PDO::PARAM_INT);
+    $delete_cart_query->bindParam(':user_id', $user_id, PDO::PARAM_INT);
 
-    if ($delete_cart_query->execute()) {
-        header('location:cart.php');
-    } else {
-        $message[] = 'Không thể xóa khỏi giỏ hàng!';
+    try {
+        $delete_cart_query->execute();
+        $message[] = 'Đã xóa khỏi giỏ hàng!';
+    } catch (PDOException $e) {
+        $message[] = 'Không thể xóa khỏi giỏ hàng! Lỗi: ' . $e->getMessage();
     }
 }
 
@@ -40,10 +27,17 @@ if (isset($_GET['delete_all'])) {
     $delete_all_cart_query = $pdo->prepare("DELETE FROM `cart` WHERE user_id = :user_id");
     $delete_all_cart_query->bindParam(':user_id', $user_id, PDO::PARAM_INT);
 
-    if ($delete_all_cart_query->execute()) {
-        header('location:cart.php');
-    } else {
-        $message[] = 'Không thể xóa tất cả!';
+    try {
+        $delete_all_cart_query->execute();
+        $message[] = 'Đã xóa hết khỏi giỏ hàng!';
+
+        // Thêm mã JavaScript để xóa tất cả sản phẩm và thông tin từ local storage
+        echo '<script>';
+        echo 'localStorage.removeItem("totalQuantity");';
+        echo 'localStorage.removeItem("grandTotal");';
+        echo '</script>';
+    } catch (PDOException $e) {
+        $message[] = 'Không thể xóa tất cả! Lỗi: ' . $e->getMessage();
     }
 }
 ?>
